move register_reads out of ReadWriteSet
